修改了language表

+++ OJ-backend/src/main/java/com/coding/oj/mapper/CommentMapper.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface CommentMapper {
     int deleteByPrimaryKey(Long id);
 
-    int insert(Comment record);
+    Long insert(Comment record);
 
     Comment selectByPrimaryKey(Long id);
